Make an admin servlet that does nothing, migrate close shielding to utils
